Added styling to login form buttons
The styling for the login form buttons has been enhanced. A new CSS class 'button-container' was added in LoginForm.scss to organize the form buttons with properties like flex, align-items, center, and gap. This class was then implemented within the LoginForm.jsx to better structure the Google and Discord login buttons.