Refactor code to move fetchUserData function from utils to ProfilePage
The fetchUserData function, initially defined in the utils.jsx file, has been moved directly inside the ProfilePage.jsx file. The function call was also added within an effect on the ProfilePage to get the currentUser data, and the updated userDiscord state variable is now displayed on the user profile page.