Add option to change username in profile page
A new feature has been introduced to allow users to change their username from the profile page. An edit pen icon has been added next to the username, clicking on which opens up a modal for changing the username. Necessary validation is put in place to handle errors and success cases.